% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxphreg.R
\name{coxphreg}
\alias{coxphreg}
\title{Cox Proportional Regression}
\usage{
coxphreg(data, lin = ~1, loglin = ~1, gcis = F)
}
\arguments{
\item{data}{A tibble/data.frame like object containing risk-sets. Variables
\code{case_id} and \code{case} are required to be contained in the data.frame.}

\item{lin}{A one sided function containing covariates in linear section of
hazard function}

\item{loglin}{A one sided function containing covariates in log-linear
section of hazard function}

\item{gcis}{A boolean (TRUE/FALSE) indicating if profile-likelihood CIs should
be calcualted.}
}
\value{
A list containing:
\itemize{
\item output:      A tibble containing paratemter estimates (est), standard errors (se), and lower/upper CIs,
\item neg2LL_diff: -2 * log-likelihood,
\item AIC:         neg2LL_diff + 2 * number of parameters,
\item hessian:     Hessian matrix o$hessian
}
}
\description{
Performs Cox proportional regression on a tibble of risk-sets
that were outputted from \code{get_rs()} function. Both a linear piece and
log-linear piece of the hazard function can be specified. The hazard function
has the general form: h(t) = h0(t) * exp(loglin) * (1 + lin).
}
