% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cum_exp.R
\name{cum_exp}
\alias{cum_exp}
\title{Calculate Cumulative Exposure}
\usage{
cum_exp(var, begin_dt, end_dt, cut_dt, lag = 0)
}
\arguments{
\item{var}{numeric representing the daily exposure value.}

\item{begin_dt}{date representing the beginning date of an exposure period.}

\item{end_dt}{date representing the end date of an exposure period.}

\item{cut_dt}{date representing the beginning date of an exposure period.}

\item{lag}{numeric representing the lag, in years, to be applied when calculating cumulative exposure.}
}
\value{
A numeric representing the cumulative, adjusting for a lag and cutoff date, for
the exposure period.
}
\description{
Calculates cumulative exposure taking into account a beginning date
end date and a cutoff date. This is useful when combining an exposure
history file and a set of risk-sets.
}
