% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_history.R
\docType{data}
\name{example_history}
\alias{example_history}
\title{Example History File for Testing}
\format{
A data frame with 9624 rows and 4 variables:
\describe{
\item{id}{unique identifier; numeric}
\item{begin_dt}{beginning date of an exposure period; character}
\item{end_dt}{beginning date of an exposure period; character}
\item{daily_exposure}{a hypothetical variable identifying daily exposure
levels to be summed to calculate a cumulative exposure; numeric}

...
}
}
\source{
{Internally Generated}
}
\usage{
example_history
}
\description{
A tibble containing example history file data to be used for testing and
demonstration of the package
}
\keyword{datasets}
