% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_rs.R
\name{gt_rs}
\alias{gt_rs}
\title{Create Risk Sets}
\usage{
gt_rs(person)
}
\arguments{
\item{person}{A data.frame like object containing one row per person with the
required demographic information.}
}
\value{
A tibble containing the risk sets for each case using age as the time-scale.
}
\description{
\code{gt_rs} creates risk-sets from a person data.frame/tibble with age as the time-scale.
It automatically matches on race, sex and date of birth (within 5 years of
the index case).
}
\details{
The person tibble must contain the variables:
\itemize{
\item id,
\item sex,
\item race,
\item dob (date),
\item pybegin (date),
\item dlo	(date),
\item case (boolean: TRUE/FALSE)
}
}
\examples{
library(dplyr)
risk_sets <- example_person \%>\%
  mutate(dob = as.Date(dob),
         pybegin= as.Date(pybegin),
         dlo = as.Date(dlo),

         sex = gender,

         case = (lung_cancer == 'TRUE')) \%>\%
  gt_rs()


}
