% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_rapid_helpers.R
\name{matrix_cells_vec}
\alias{matrix_cells_vec}
\title{Helpers for the \code{get_table_rapid()} function}
\usage{
matrix_cells_vec(data, time_break, matrix_obj)
}
\arguments{
\item{data}{data from get_table_rapid()}

\item{time_break}{number of years definedy by user}

\item{matrix_obj}{a matrix like object}
}
\description{
Helpers for the \code{get_table_rapid()} function
}
